+++ src/Databricks/Databricks/ChangeLog.md
@@ -18,6 +18,7 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Updated the Az Databricks cmdlets to 2024-05-01 api version.
 
 ## Version 1.7.2
 * Introduced secrets detection feature to safeguard sensitive data.